Fix workflow -Werror type-limits

+++ src/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: danalvar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/15 18:08:31 by danalvar          #+#    #+#             */
-/*   Updated: 2025/01/29 14:59:05 by danalvar         ###   ########.fr       */
+/*   Updated: 2025/01/29 15:35:03 by danalvar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	len;
+	int	len;
 
 	len = ft_strlen(s);
 	while (len >= 0)
